use smaller ints for x,u

+++ print.ino
@@ -1,4 +1,4 @@
-void prints(int x, int y, const char *message, bool inv) {
+void prints(int8_t x, uint8_t y, const char *message, bool inv) {
   font3x5.setCursor( x*4+1, y*7 );
   if(inv == 0){
     font3x5.setTextColor(WHITE);
@@ -10,7 +10,7 @@ void prints(int x, int y, const char *message, bool inv) {
   }
 }
 
-void prints(int x, int y, int d, bool inv){
+void prints(int8_t x, uint8_t y, int d, bool inv){
   font3x5.setCursor( x*4+1, y*7 );
   if(inv == 0){
     font3x5.setTextColor(WHITE);
